Add files via upload
Users must first enter the predefined password, which will be read by the keypad, and if the password is correct, the servo motor will turn on and turn the door lock; if the password is incorrect, the servo motor will not turn on, and we must retype the password which have been predefined in the programme. Passwords can be updated as needed by the user.
This is a low-cost application with simple to programme code. And the password can be updated as needed by the user. In comparison to traditional door locks, the technology requires less power to operate. Furthermore, users are no longer required to carry physical keys. 
As long as the password is not shared with anyone, this application can provide perfect security. Users can introduce constraints to the password entry process to make it safer. Also add a fingerprint scanner and a display model to show messages such as whether the door is locked or not. Instead of using a keypad module, can employ smart cards to open doors with the same model. In the future, the proposed system can employ a power supply adaptor instead of a battery for power transfer. It is cost effective because we are using affordable component costs. Also, with the help of the library, it is extremely simple to develop code, and anyone may use this model for security purposes. Because the existing door lock system has an issue with expensive and irreplaceable components, one may use an Arduino-based door lock system to solve this problem because it is simple to install and remove. Therefore, the "Password-based door lock system using Arduino" is a time-saving programmable module that will assist us in providing excellent security.